Slice 4.2: move settings panels to new /admin/settings tab
Admin IA cleanup. Both settings panels ("Brief delivery" + "Announcement
authors") lived under the Civic Briefs tab because that's where the
settings plumbing was introduced. But author management and brief
recipient config aren't part of the brief review workflow — they're
hub-wide configuration.

- New pages/AdminSettings.tsx + .css: dedicated page with both panels,
  their state, and save handlers. Renders under AdminTabs at
  /admin/settings.
- AdminTabs.tsx: added a third tab (Proposals · Civic Briefs · Settings).
- pages/AdminBriefs.tsx: removed all settings state, handlers, and JSX.
  Dropped unused imports (adminGetSettings, adminPatchSettings,
  AnnouncementAuthor). The page is purely the brief review queue again.
- pages/AdminBriefs.css: removed panel/row styles (relocated to
  AdminSettings.css).
- App.tsx: new /admin/settings route.

Renamed panel heading "Delivery settings" → "Brief delivery" for
specificity ("delivery" alone would collide with announcement delivery
if that ever becomes a thing).

No backend changes — /admin/settings API + hub_settings table
unchanged. Pure UI reorganization.
